Setup for map component started, alterations to location and trips needs addressed
