+++ ui/src/Components/map.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import GoogleMapReact from 'google-map-react';
+
+interface mapProps {
+    location: any,
+    zoomLevel: any
+}; 
+
+function Map(props: mapProps) {
+  return (
+    <>
+    
+    </>
+  )
+}
+
+export default Map;
