+++ ui/src/Pages/home/home.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Home() {
+  return (
+    <>
+        
+    </>
+  )
+}
+
+export default Home;
