+++ ui/src/Pages/index.tsx
@@ -1,1 +1,1 @@
-export { default as Root } from "./root";
+<></>
