+++ ui/src/types/location.ts
@@ -1,0 +1,5 @@
+export type Location = {
+    address: string,
+    latitude: number,
+    longitude: number
+};
